+++ test/dummy/db/migrate/20260903000006_add_last_used_at_to_recording_studio_oauth_access_tokens.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddLastUsedAtToRecordingStudioOauthAccessTokens < ActiveRecord::Migration[8.1]
+  def change
+    add_column :recording_studio_oauth_access_tokens, :last_used_at, :datetime
+  end
+end